DB Connection and Functionality

+++ ARJAPIADo/Models/Employee.cs
@@ -2,7 +2,7 @@
 {
     public class Employee
     {
-        public int EmployeeId { get; set; }
+        public long EmployeeId { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string PhoneNo { get; set; }
